+++ CHANGELOG.md
@@ -6,8 +6,8 @@
     carousel rotations.
 - 🌈 order book component. type `^` or `book` to see an order book for the
     active symbol.
-- ⌚ polling now available for all components. Use command `poll` or set
-    config component key `pollMs` to a value greater than 10.
+- ⌚ polling now available for all components. Use prefix-command `poll` or
+    set config component key `pollMs` to a value greater than 10.
 
 ## v0.0.11
 - filepath fixes